Se agrega controllers, dto y middleware

+++ src/main/java/com/coderhouse/java/persistences/repositories/InvoiceDetailRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface InvoiceDetailRepository extends JpaRepository<InvoiceDetail, Integer> {
+public interface InvoiceDetailRepository extends JpaRepository<InvoiceDetail, Long> {
     List<InvoiceDetail> findAllByInvoiceId(Long invoiceId);
 }
